backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
celery[redis]==5.4.0
redis==5.2.1
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
hypothesis==6.122.3
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
aiofiles==24.1.0
python-multipart==0.0.20
qdrant-client==1.12.1
tiktoken==0.8.0
reportlab==4.2.5
ebooklib==0.18
structlog==24.4.0
tenacity==9.0.0
asyncpg==0.30.0
aiosqlite==0.20.0
greenlet==3.1.1
sentence-transformers>=2.2.0
torch>=2.0.0
